First step.
It starts walking but there's a lot of work pending to be done. 

I'll be writting comments and reorganizing the whole code since it's
more like a POC than a formal project now.

This commit is just the code I will upload makefiles, a compiled binary
and other stuff like simple uml diagrams thanks to cpp2dia ;), it saved
some work here.